Rename example_loader -> dataset_loader
