+++ pyvista/examples/_dataset_loader.py
